+++ pkg/server/commands/list_objects.go
@@ -321,7 +321,11 @@ func (q *ListObjectsQuery) evaluate(
 			}
 		}
 
-		reverseExpandResultsChan := make(chan *reverseexpand.ReverseExpandResult, 1)
+		var bufferSize uint32
+		cappedMaxResults := uint32(math.Min(float64(maxResults), 1000)) // cap max results at 1000
+		bufferSize = uint32(math.Max(float64(cappedMaxResults/10), 10)) // 10% of max results, but make it at least 10
+
+		reverseExpandResultsChan := make(chan *reverseexpand.ReverseExpandResult, bufferSize)
 		objectsFound := atomic.Uint32{}
 
 		ds := storagewrappers.NewRequestStorageWrapperWithCache(